Began to add UI, but in doing so I broke other rendering!

+++ core/src/com/joelallison/level/Map.java
@@ -10,7 +10,7 @@ public class Map {
 
     public ArrayList<String[][]> mapsToCombine;
 
-    public static float[][] genNoiseMap (long seed, Vector2 Dimensions, float xOffset, float yOffset, float scale, int octaves, float persistence, float lacunarity, int wrapFactor, boolean invertWrap) {
+    public static float[][] genNoiseMap (long seed, Vector2 Dimensions, int xOffset, int yOffset, float scale, int octaves, float persistence, float lacunarity, int wrapFactor, boolean invertWrap) {
         //greater scale zooms in, halved scale from normal could be used for map data. I think scale of 4 is best for most stuff
         //higher octaves adds more detail to the noise, but 2 is the best option for this [in most cases]
         //higher persistence makes the values tend towards higher values
@@ -49,7 +49,7 @@ public class Map {
                     minNoiseHeight = noiseHeight;
                 }
 
-                noiseMap[x-(int)xOffset][y-(int)yOffset] = noiseHeight;
+                noiseMap[x-xOffset][y-yOffset] = noiseHeight;
             }
         }
